Convert to use 24-hour time for a full circuit
I like the look of this watchface, but it seemed weird to me that it's 12-hour time and thus loops around at noon again. For my own purposes at least I'd prefer 24-hour, and one of the signs that this is a great example watchface is how easy it was for me to lazily hack that!